backend/requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
